+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Support switch consortium active affiliation. Refs STCOR-690.
 * Use tenant query param for password reset. Refs STCOR-724.
 * Improve login error message when back-end not ready. Refs STCOR-723.
+* Replace `align-items: start` with `flex-start`; it is more widely supported. Refs STCOR-722.
 
 ## [9.0.0](https://github.com/folio-org/stripes-core/tree/v9.0.0) (2023-01-30)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v8.3.0...v9.0.0)