+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for stripes-core
 
-## 9.1.0 IN PROGRESS
+## 10.0.0 IN PROGRESS
 
 * Close app-context menu when its wrapper is undefined, which it will be when switching among apps. Fixes STCOR-680.
 * Move `miragejs` to dev-deps; remove `clean` script and `rimraf`. Refs STCOR-668, STCOR-681.